Refactored: Global configuration to use configuration object.
