+++ lib/ddtrace/configuration/option_set.rb
@@ -1,0 +1,6 @@
+module Datadog
+  module Configuration
+    class OptionSet < Hash
+    end
+  end
+end
