--- lib/ddtrace/contrib/configuration/option_definition_set.rb
@@ -1,20 +1,0 @@
-require 'ddtrace/configuration/resolver'
-
-module Datadog
-  module Contrib
-    module Configuration
-      # Represents a set of configuration option definitions for an integration
-      class OptionDefinitionSet < Hash
-        def dependency_order
-          Datadog::Configuration::Resolver.new(dependency_graph).call
-        end
-
-        def dependency_graph
-          each_with_object({}) do |(name, option), graph|
-            graph[name] = option.depends_on
-          end
-        end
-      end
-    end
-  end
-end
