--- lib/ddtrace/contrib/configuration/option_set.rb
@@ -1,8 +1,0 @@
-module Datadog
-  module Contrib
-    module Configuration
-      class OptionSet < Hash
-      end
-    end
-  end
-end
