+++ spec/ddtrace/configuration/option_definition_set_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 require 'ddtrace'
 
-RSpec.describe Datadog::Contrib::Configuration::OptionDefinitionSet do
+RSpec.describe Datadog::Configuration::OptionDefinitionSet do
   subject(:set) { described_class.new }
 
   it { is_expected.to be_a_kind_of(Hash) }
@@ -10,17 +10,17 @@ RSpec.describe Datadog::Contrib::Configuration::OptionDefinitionSet do
   shared_context 'dependent option set' do
     before(:each) do
       set[:foo] = instance_double(
-        Datadog::Contrib::Configuration::OptionDefinition,
+        Datadog::Configuration::OptionDefinition,
         depends_on: [:bar]
       )
 
       set[:bar] = instance_double(
-        Datadog::Contrib::Configuration::OptionDefinition,
+        Datadog::Configuration::OptionDefinition,
         depends_on: [:baz]
       )
 
       set[:baz] = instance_double(
-        Datadog::Contrib::Configuration::OptionDefinition,
+        Datadog::Configuration::OptionDefinition,
         depends_on: []
       )
     end
@@ -30,10 +30,10 @@ RSpec.describe Datadog::Contrib::Configuration::OptionDefinitionSet do
     subject(:dependency_order) { set.dependency_order }
 
     context 'when invoked' do
-      let(:resolver) { instance_double(Datadog::Configuration::Resolver) }
+      let(:resolver) { instance_double(Datadog::Configuration::DependencyResolver) }
 
       it do
-        expect(Datadog::Configuration::Resolver).to receive(:new)
+        expect(Datadog::Configuration::DependencyResolver).to receive(:new)
           .with(a_kind_of(Hash))
           .and_return(resolver)
         expect(resolver).to receive(:call)
