+++ spec/ddtrace/configuration/option_set_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 require 'ddtrace'
 
-RSpec.describe Datadog::Contrib::Configuration::OptionSet do
+RSpec.describe Datadog::Configuration::OptionSet do
   subject(:set) { described_class.new }
 
   it { is_expected.to be_a_kind_of(Hash) }
