+++ spec/ddtrace/configuration/option_spec.rb
@@ -2,11 +2,11 @@ require 'spec_helper'
 
 require 'ddtrace'
 
-RSpec.describe Datadog::Contrib::Configuration::Option do
+RSpec.describe Datadog::Configuration::Option do
   subject(:option) { described_class.new(definition, context) }
   let(:definition) do
     instance_double(
-      Datadog::Contrib::Configuration::OptionDefinition,
+      Datadog::Configuration::OptionDefinition,
       default_value: default_value,
       setter: setter
     )
