+++ spec/ddtrace/contrib/configuration/settings_spec.rb
@@ -5,7 +5,7 @@ require 'ddtrace'
 RSpec.describe Datadog::Contrib::Configuration::Settings do
   subject(:settings) { described_class.new }
 
-  it { is_expected.to be_a_kind_of(Datadog::Contrib::Configuration::Options) }
+  it { is_expected.to be_a_kind_of(Datadog::Configuration::Options) }
 
   describe '#options' do
     subject(:options) { settings.options }
